# docs/openApi.yml
openapi: 3.2.0
info:
  title: T-Shirt Store-OpenAPI
  version: 1.0.0
  description: |-
    This is first version of the contract for T-Shirt Store-OpenAPI. You can find out more about the patterns followed here https://github.com/CAMIANAIS/T-Shirt-Store-API/blob/main/docs/openApi_Patterns.md

    ## Authentication

    All endpoints require Bearer token authentication via JWT, except for the following public endpoints:
    - `POST /auth/signup` — Registration (can't have token yet)
    - `POST /auth/signin` — Login (can't have token yet)
    - `POST /auth/forgotpassword` — Password reset request (can't have token yet)
    - `POST /auth/resetpassword` — Password reset completion (can't have token yet)
    - `GET /products`, `GET /products/{productId}`, `GET /products/{productId}/variants`, `GET /products/{productId}/images`, `GET /categories` — Product/category browsing doesn't require authentication
    - `POST /webhooks/stripe` — Called by Stripe, not a user; authenticated via Stripe-Signature header instead of a bearer token

    Authentication failures return 401 (Authentication credentials invalid or missing) or 403 (insufficient permissions). The server reads the JWT role and automatically scopes results (regular users see only their own data; managers see all data).
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Product:
      required:
        - name
        - description
        - status
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: "summer-tshirt"
        description:
          type: string
          example: "long sleeves, 100% cotton"
        status:
          type: string
          enum:
            - active
            - inactive
            - discontinued
          example: "active"
        category:
          $ref: "#/components/schemas/Category"
        createdAt:
          type: string
          format: date-time
          example: "1985-04-12T23:20:50.52Z"
        updatedAt:
          type: string
          format: date-time
          example: "1985-04-12T23:20:50.52Z"
        deletedAt:
          type: string
          format: date-time
          example: "1985-04-12T23:20:50.52Z"
    Category:
      required:
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "tshirts"

        createdAt:
          type: string
          format: date-time
          example: "1985-04-12T23:20:50.52Z"
        updatedAt:
          type: string
          format: date-time
          example: "1985-04-12T23:20:50.52Z"
        deletedAt:
          type: string
          format: date-time
          example: "1985-04-12T23:20:50.52Z"
...
    ProductVariantInput:
      required:
        - size
        - color
        - stockQuantity
        - skuCode
      type: object
      properties:
        size:
          type: string
          enum:
            - "S"
            - "M"
            - "L"
            - "XL"
            - "XXL"
          example: "S"
        color:
          type: string
          example: "red"
        stockQuantity:
          type: integer
          format: int64
          example: 1

        skuCode:
          type: string
          pattern: "^[a-zA-Z0-9]+$"
          example: "TSH-BLU-MD"
        status:
          type: string
          enum:
            - active
            - inactive
            - discontinued
...
    ProductImage:
      required:
        - bucketName
        - filename
        - extension
        - displayOrder
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        bucketName:
          type: string
          example: "bucketNameHere"
        filename:
          type: string
          example: "fileNameHere"
        extension:
          type: string
          enum:
            - "jpg"
            - "png"
            - "webp"
            - "gif"
          example: "jpg"
        displayOrder:
          type: integer
          format: int64
          example: 1
        altText:
          type: string
          example: "Product front view"
        createdAt:
          type: string
          format: date-time
          example: "1985-04-12T23:20:50.52Z"
        updatedAt:
          type: string
          format: date-time
          example: "1985-04-12T23:20:50.52Z"
        product:
          $ref: "#/components/schemas/Product"
    ProductImageInput:
      required:
        - file
      type: object
      properties:
        file:
          type: string
          format: binary
          description: "Raw image file bytes. Supported formats: jpg, png, webp, gif"
        displayOrder:
          type: integer
          format: int64
          example: 1
        altText:
          type: string
          example: "Product front view"
    ProductInput:
      required:
        - name
        - description
        - status
      type: object
      properties:
        name:
          type: string
          example: "summer-tshirt"
        description:
          type: string
          example: "long sleeves, 100% cotton"
        status:
          type: string
          enum:
            - active
            - inactive
            - discontinued
          example: "active"
        categoryId:
          type: integer
          description: Set a category for this product
          example: 1
        variants:
          type: array
          description: Optional variants to create alongside the product in one call
          items:
            $ref: "#/components/schemas/ProductVariantInput"
    ProductUpdateInput:
      type: object
      properties:
        name:
          type: string
          example: "summer-tshirt"
        description:
          type: string
          example: "long sleeves, 100% cotton"
        status:
          type: string
          enum:
            - active
            - inactive
            - discontinued
          example: "active"
        categoryId:
          type: integer
          description: Set a category for this product
          example: 1
...
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: "Product not found"
        error:
          type: string
          example: "Not Found"
...
security:
  - bearerAuth: []
paths:
  /products/{productId}:
    get:
      tags:
        - product
      summary: Find product by identifier
      description: Returns a single product.
      operationId: getProductById
      parameters:
        - name: productId
          in: path
          description: ID of product to return
          required: true
          schema:
            type: integer
            format: int64
      security: []
      responses:
        "200":
          description: Product successfully returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
            application/xml:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid productId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "default":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags:
        - product
      summary: Update product
      description: Updates any field of a product
      operationId: updateProductItem
      parameters:
        - name: productId
          in: path
          description: ID of the product to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update the product
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdateInput"
          application/xml:
            schema:
              $ref: "#/components/schemas/ProductUpdateInput"
        required: true
      responses:
        "200":
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
            application/xml:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid productId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication credentials invalid or missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "default":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - product
      summary: Soft Delete a product
      description: Removes a product using a logical delete.
      operationId: removeProduct
      parameters:
        - name: productId
          in: path
          description: ID of product to remove
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Item removed successfully
        "400":
          description: Invalid productId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication credentials invalid or missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "default":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /products:
    get:
      tags:
        - product
      summary: Get all products
      description: Returns a list of products
      operationId: getProducts
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
            description: Filter products by category id
            example: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 50
            description: Show products considering this limit
            example: 1
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            description: Show products considering this offset
            example: 1
      security: []
      responses:
        "200":
          description: Products successfully returned
          headers:
            X-TotalCount:
              description: Total number of products matching the applied filters
              schema:
                type: integer
          content:
            application/json:
              schema:
                {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Product" },
                }
            application/xml:
              schema:
                {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Product" },
                }
        "400":
          description: Invalid category or pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "default":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags:
        - product
      summary: Add a new product.
      description: Add a new product.
      operationId: addProduct
      requestBody:
        description: Create a new product in the store
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
          application/xml:
            schema:
              $ref: "#/components/schemas/ProductInput"
        required: true
      responses:
        "201":
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
            application/xml:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication credentials invalid or missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "default":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /products/{productId}/images:
    get:
      tags:
        - product
      summary: Find images by product Id.
      description: Returns images by product Id.
      operationId: getProductImages
      parameters:
        - name: productId
          in: path
          description: ID of product to return
          required: true
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 50
            description: Show productImages considering this limit
            example: 1
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            description: Show productImages considering this offset
            example: 1
      security: []
      responses:
        "200":
          description: ProductImages successfully returned
          headers:
            X-TotalCount:
              description: Total number of productImages matching the applied filters
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductImage"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductImage"
        "400":
          description: Invalid productId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "default":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - product
      summary: Add a new productImage.
      description: Add a new productImage.
      operationId: addProductImage
      parameters:
        - name: productId
          in: path
          description: ID of product to return
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Create a new productImage in the store
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ProductImageInput"
        required: true
      responses:
        "201":
          description: ProductImage created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductImage"
        "400":
          description: Missing file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication credentials invalid or missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "default":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /products/{productId}/activate:
    post:
      tags:
        - product
      summary: Activate a product
      description: |
        Sets the product's status to `active`. Idempotent: calling this on an already-active
        product just returns it unchanged, no error. Manager-only.
      operationId: activateProduct
      parameters:
        - name: productId
          in: path
          description: ID of the product to activate
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Product activated successfully (or already active)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid productId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication credentials invalid or missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Cannot activate a discontinued product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "default":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /products/{productId}/deactivate:
    post:
      tags:
        - product
      summary: Deactivate a product
      description: |
        Sets the product's status to `inactive`. Idempotent: calling this on an already-inactive
        product just returns it unchanged, no error. Manager-only.
      operationId: deactivateProduct
      parameters:
        - name: productId
          in: path
          description: ID of the product to deactivate
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Product deactivated successfully (or already inactive)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid productId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication credentials invalid or missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Cannot deactivate a discontinued product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "default":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /products/{productId}/likes:
    post:
      tags:
        - product
      summary: Like a product
      description: |
        Add the authenticated user to the product's likes.

        **Idempotent:** calling this multiple times has the same effect as calling it once.
        If the user already likes this product, returns 204 (success) with no change.
      operationId: likeProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 42
      responses:
        "204":
          description: Product liked successfully (or already liked by this user)
        "401":
          description: Authentication credentials invalid or missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "default":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - product
      summary: Unlike a product
      description: |
        Remove the authenticated user from the product's likes.

        **Idempotent:** calling this multiple times has the same effect as calling it once.
        If the user does not like this product, returns 204 (success) with no change.
      operationId: unlikeProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 42
      responses:
        "204":
          description: Product unliked successfully (or was not liked by this user)
        "401":
          description: Authentication credentials invalid or missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "default":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /products/{productId}/paymentLink:
    post:
      tags:
        - product
      summary: Generate payment link for single product
      description: Create a shareable Stripe Payment Link for direct product purchase
      parameters:
        - name: productId
          in: path
          description: ID of product to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "201":
          description: paymentLink created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  paymentLink:
                    type: string
                    example: "https://buy.stripe.com/test_..."
                  expiresAt:
                    type: string
                    format: date-time
        "401":
          description: Authentication credentials invalid or missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "default":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...